defai-tele/requirements.txt
web3==6.11.1
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-dotenv==1.0.0
tensorflow==2.14.0
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
pydantic==2.5.2
python-binance==1.0.19
aiohttp==3.9.1
pytest==7.4.3
black==23.11.0
isort==5.12.0 